scanblog/requirements.txt
django>=1.3.1
psycopg2==2.4.1
-e hg+http://bitbucket.org/ubernostrum/django-registration#egg=django-registration
# Require celery 2.2 because of backwards incompatibility introduced in 2.3 --
# upgrade when resolved.
Celery<2.3
django-celery<2.3
# Require python-dateutil 1.5 due to this issue: https://groups.google.com/group/celery-users/browse_thread/thread/08f754dbba35d1f4/b2587c10d65c64e3?lnk=raot
python-dateutil==1.5
pyPdf
PIL
amqplib
fabric
django-bcrypt
-e git+https://github.com/ahupp/python-magic#egg=magic
lettuce
django_compressor
sorl-thumbnail
python-memcached
-e git+https://github.com/jtauber/django-notification#egg=django-notification
django-pagination
django-urlcrypt
South
pyyaml
django-debug-toolbar
reportlab
